Handle empty non-last write and fix out of bounds write
An empty write would cause a `\n` to be emitted if `last_was_cr` was true,
which is incorrect because the next chunk might contain that `\n`.

Because we relaxed the size check to 2n for non-last writes we can no longer
just write an extra `\n` unconditionally at the end of the output buffer when
we detect that the input ends with a `\r`.