Split `Normalize` into `Normalize` and `NormalizeChunk`
`NormalizeChunk` is now the low-level trait that consumers shouldn't need
to interact with unless they want manually normalize chunks of data.

The reason to do this is that `normalize` method with the &mut [MaybeUninit<u8>]
is easy to misuse, and it's better to have a higher-level trait that
provides safe methods for normalizing entire buffers. Or the higher-level
traits to create stream wrappers that normalize data on the fly.